foxxy-bot/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.2.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.9
docopt==0.6.2
frozenlist==1.3.0
idna==3.3
MouseInfo==0.1.3
multidict==5.2.0
nextcord==2.0.0a7
oauthlib==3.1.1
pipreqs==0.4.11
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dotenv==0.19.2
pytweening==1.0.4
requests==2.26.0
requests-oauthlib==1.3.0
tweepy==4.4.0
typing-extensions==3.10.0.2
urllib3==1.26.7
yarg==0.1.9
yarl==1.7.2
